refactor template.go out of terraform to be used as common code
Now that we are adding a govmomi based implementation as well, the
script templates as they stand today should be used by both govmomi
and terraform implementations. Thus refactoring the template.go to
be consumable with both implementations

Resolves #57
